Modify Reminder 3 + Correction
initialization of the values of the cells you wish to modify.
Displaying the date (modify or create a task)
Fixed bug for selecting a task of type "other".
Fixed the bug when deleting a task (core data bug when restarting the application), it was necessary to try try save context